feat(maestro): split reader.py into a package; address #46
Closes #46.

## Module split

reader.py (2200 lines, 9 responsibilities) -> reader/ package:
  _skill.py         low-level SKILL helpers
  _parse_skill.py   s-expr / alist / sev-output parsers
  _parse_sdb.py     maestro.sdb XML parsers
  _compact.py       snapshot reshape helpers
  remote_io.py      scp-backed file read
  session.py        read_session_info + focus resolution
  probes.py         read_config/env/variables/outputs/corners/status
  runs.py           read_results, history enumeration, latest-log parse
  snapshot.py       snapshot + snapshot_to_dir

Public API surface is unchanged -- external callers go through
`from virtuoso_bridge.virtuoso.maestro import X`, which still resolves
via the package __init__.

## Issue #46 enhancements

1. SKILL-first scratch_root detection
   New `detect_scratch_root_via_skill` uses `asiGetAnalogRunDir`, the
   authoritative Cadence API.  `snapshot_to_dir` tries SKILL first, falls
   back to the sdb regex heuristic on failure.  Works on fresh
   (un-simulated) sessions.

2. scratch_runs[] instead of single-run scratch_files
   latest_history now emits one entry per run with:
     run_id, hist_tag, corner, corner_process, temperature,
     netlist_scs, design_variables, model_files, spectre_out,
     psf_dir, psf_files (analysis short-name -> path)
   One batched SKILL call reads all runs' .modelFiles / input.scs /
   psf dir listings.

3. spectre_tail only on error
   On successful runs the tail is license-banner noise; drop it unless
   spectre_errors_count > 0.

4. analysis tag on computed outputs
   Derived from the SKILL expression: `?result "NAME"` wins, else
   VF/IF -> ac, VT/IT -> tran, VDC/IDC/VS/IS -> dc, VOP -> dcOp.

## Breaking API changes

- snap["results"]  -> snap["output_values"]
- snap["outputs"]  -> snap["output_defs"]
- snapshot(include_results=...)  -> snapshot(include_output_values=...)
- read_latest_history top-level run_id / hist_tag / scratch_files
  -> scratch_runs[]
- read_corners(client, session, sdb_path=...)
  -> read_corners(client, sdb_path, ...)   (sdb_path now required)

## Minor wins

- snapshot() auto-creates a shared sdb temp cache so sub-readers only
  fetch maestro.sdb once (was 4-5 scp per snapshot).
- `virtuoso-bridge windows` highlights the SKILL-focused window with a
  top-line summary + `*` prefix + ANSI bold.
- Removed dead fallback in read_corners (TypeError on untriggered path).
